增加字典管理功能

+++ source/server/types/Dist.go
@@ -1,9 +1,9 @@
 package types
 
 type Dist struct {
-	Id        int    `json:"id"`
+	Id        int64  `json:"id"`
 	Type      string `json:"type"`
 	DistKey   string `json:"distKey"`
 	DistValue string `json:"distValue"`
-	Sort      int    `json:"sort"`
+	Sort      int64  `json:"sort"`
 }
